+++ src/sps/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.x.x"
+__version__ = "0.2.0"